Continuation of profile capabilty as well as forum (Community Tab) functionality implemented
Users can now choose a jpeg or png file to upload as their profile pic, it encodes to base64 to store the image in firestore, this somewhat degrades quality, but its miniscule. The community page is now set up, you can post and it will log the timestamp as well as username and profile pic, I currently have it where you can only post if you are logged into the wonderful NoTreble platform, if not it simply wont let you post, so if you try it out and dont see anywhere to post, thats why, you gotta make an account baby :) Otherwise, thats bout it for now. P.S i assume someone is working on settings/Profile tab in the taskbar, If you wanna see profile as of now ive just been forceably accessing the page with /Profile at the end of the local host site, after logging in of course, so if you wanna add a pic, you gotta do that :P